requirements.txt
pdfplumber==0.10.3
python-dotenv==1.0.0
cryptography
matplotlib
